Refactor
Create factory instead of dedicated class because
because dedicated class actually was a factory method
that had reusable behaviour.
